refactor: latest.js
